gdk/x11: Move GdkX11GLContext definitions into the source
There's no need to have the instance and class structures in the private
header: nothing accesses them directly.
